requirements.txt
streamlit
pdfplumber
google-generativeai
python-dotenv
beautifulsoup4
requests
PyMuPDF
groq
